remove CompactionTaskFactory
