+++ server/src/main/java/org/apache/iotdb/db/engine/compaction/constant/InnerCompactionStrategy.java
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -16,64 +16,65 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.engine.compaction.cross;
+
+package org.apache.iotdb.db.engine.compaction.constant;
 
 import org.apache.iotdb.db.engine.compaction.CompactionTaskManager;
-import org.apache.iotdb.db.engine.compaction.cross.rewrite.RewriteCrossSpaceCompactionSelector;
-import org.apache.iotdb.db.engine.compaction.cross.rewrite.task.RewriteCrossSpaceCompactionTask;
+import org.apache.iotdb.db.engine.compaction.inner.AbstractInnerSpaceCompactionSelector;
+import org.apache.iotdb.db.engine.compaction.inner.AbstractInnerSpaceCompactionTask;
+import org.apache.iotdb.db.engine.compaction.inner.sizetiered.SizeTieredCompactionSelector;
+import org.apache.iotdb.db.engine.compaction.inner.sizetiered.SizeTieredCompactionTask;
 import org.apache.iotdb.db.engine.storagegroup.TsFileManager;
 import org.apache.iotdb.db.engine.storagegroup.TsFileResource;
 
 import java.util.List;
 
-public enum CrossCompactionStrategy {
-  REWRITE_COMPACTION;
+public enum InnerCompactionStrategy {
+  SIZE_TIERED_COMPACTION;
 
-  public static CrossCompactionStrategy getCrossCompactionStrategy(String name) {
-    if ("REWRITE_COMPACTION".equalsIgnoreCase(name)) {
-      return REWRITE_COMPACTION;
+  public static InnerCompactionStrategy getInnerCompactionStrategy(String name) {
+    if ("SIZE_TIERED_COMPACTION".equalsIgnoreCase(name)) {
+      return SIZE_TIERED_COMPACTION;
     }
-    throw new RuntimeException("Illegal Cross Compaction Strategy " + name);
+    throw new RuntimeException("Illegal Compaction Strategy " + name);
   }
 
-  public AbstractCrossSpaceCompactionTask getCompactionTask(
+  public AbstractInnerSpaceCompactionTask getCompactionTask(
       String logicalStorageGroupName,
-      String virtualStorageGroupName,
-      long timePartitionId,
+      String virtualStorageGroup,
+      long timePartition,
       TsFileManager tsFileManager,
-      List<TsFileResource> selectedSeqTsFileResourceList,
-      List<TsFileResource> selectedUnSeqTsFileResourceList) {
+      List<TsFileResource> selectedTsFileResourceList,
+      boolean sequence) {
     switch (this) {
-      case REWRITE_COMPACTION:
+      case SIZE_TIERED_COMPACTION:
       default:
-        return new RewriteCrossSpaceCompactionTask(
+        return new SizeTieredCompactionTask(
             logicalStorageGroupName,
-            virtualStorageGroupName,
-            timePartitionId,
+            virtualStorageGroup,
+            timePartition,
             tsFileManager,
-            selectedSeqTsFileResourceList,
-            selectedUnSeqTsFileResourceList,
+            selectedTsFileResourceList,
+            sequence,
             CompactionTaskManager.currentTaskNum);
     }
   }
 
-  public AbstractCrossSpaceCompactionSelector getCompactionSelector(
+  public AbstractInnerSpaceCompactionSelector getCompactionSelector(
       String logicalStorageGroupName,
-      String virtualGroupId,
-      String storageGroupDir,
+      String virtualStorageGroupName,
       long timePartition,
       TsFileManager tsFileManager,
-      CrossSpaceCompactionTaskFactory taskFactory) {
+      boolean sequence) {
     switch (this) {
-      case REWRITE_COMPACTION:
+      case SIZE_TIERED_COMPACTION:
       default:
-        return new RewriteCrossSpaceCompactionSelector(
+        return new SizeTieredCompactionSelector(
             logicalStorageGroupName,
-            virtualGroupId,
-            storageGroupDir,
+            virtualStorageGroupName,
             timePartition,
             tsFileManager,
-            taskFactory);
+            sequence);
     }
   }
 }
